Envio de alerta de tempo fixo Refatoração de tempo de estacionamento

+++ paquimetro/src/main/java/com/github/rafaelfernandes/parquimetro/parking/controller/request/FixTime.java
@@ -1,6 +1,6 @@
 package com.github.rafaelfernandes.parquimetro.parking.controller.request;
 
 public record FixTime(
-        Integer duration
+        Long duration
 ) {
 }
